fixed fsrs bugs

+++ Flashcards/fsrs.h
@@ -2,7 +2,7 @@
 #define FSRS_H
 
 double fsrs_retrievability(double t, double s);
-long fsrs_interval(double r, double s);
+double fsrs_interval(double r, double s);
 double fsrs_s_0(int g);
 double fsrs_stability(double r, double s, double d, int g);
 double fsrs_d_0(int g);
